Finally think I got it
I had written a couple different projects, and the one that worked didn't save when my laptop went to sleep while I was gone.  I tried starting that one over and couldn't get it to work properly. This morning though the coffee was good, and I thought of a simpler project that combined pretty much all of our previous assignments.  Hope you are having a great day!